refactor: profit 자료형 double에서 BigDecimal로 변경

+++ src/main/java/domain/result/DealerProfit.java
@@ -1,3 +1,5 @@
 package domain.result;
 
-public record DealerProfit(double profit) {}
+import java.math.BigDecimal;
+
+public record DealerProfit(BigDecimal profit) {}
