+++ src/main/java/domain/result/UserProfit.java
@@ -1,3 +1,5 @@
 package domain.result;
 
-public record UserProfit(RoundBetInfo roundBetInfo, GameResult gameResult, double profit) {}
+import java.math.BigDecimal;
+
+public record UserProfit(RoundBetInfo roundBetInfo, GameResult gameResult, BigDecimal profit) {}
