+++ src/main/java/strategy/BettingRule.java
@@ -2,7 +2,8 @@ package strategy;
 
 import domain.result.GameResult;
 import domain.result.RoundBetInfo;
+import java.math.BigDecimal;
 
 public interface BettingRule {
-    double calculateBetAmount(RoundBetInfo roundBetInfo, GameResult gameResult);
+    BigDecimal calculateBetAmount(RoundBetInfo roundBetInfo, GameResult gameResult);
 }
